Handle duplicate "Vary" headers in response
All calls were crashing because the test enpoint is sending multiple
"Vary" headers, instead of concatenating them.

This change makes the headers unique, so all those ToDictionary() calls
don't throw exceptions.

fixes:
 https://github.com/CyberSource/cybersource-rest-client-dotnetstandard/issues/94

" a sender MUST NOT generate multiple header fields with
the same field name in a message unless either the entire
field value for that header field is defined as
 a comma-separated list, [i.e., # (values)] or
 the header field is a well-known exception."